move list-services to it's own package
